backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
opencv-python>=4.9.0
tensorflow>=2.16.0
Pillow>=10.0.0
